Gracefully handle giant PDF images
If we ask poppler to render a PDF at a particularly high resolution,
say, native printer resolution, we can run into memory allocation
failures.  In theory, we could solve this by only rendering a portion of
the image but that would break some of our abstractions.  Instead,
downsize the image until we get something that fits in memory.